Update new_rental_notifier.py
Move the code to a NewRentalNotifier class with the intention of keeping track of a series of log messages as the searching process is performed for each city. After each city is searched all the log information is gathered and printed to a log.txt. This log file replaces the empty file that was created with a time stamp, which had only been implemented to ensure that the process was actually running after scheduling with Task Scheduler.